Expose `settled` helper function.
* Rename `wait` to `settled`
* Ensure existing `import wait from 'ember-test-helpers/wait'` code works properly
* Export `settled` as public API from the main `index.js`
* Move existing `wait` tests into `legacy-0-6-x` folder

Primary reasoning for the rename from `wait` to `settled`:

```js

test('can do xyz', async function(assert) {
  await this.render(hbs`{{some-thing}}`);
  await click('.some-thing');
  await wait(); // <- WAT?!?!!?
});
```

This looks much better:

```js
test('can do xyz', async function(assert) {
  await this.render(hbs`{{some-thing}}`);
  await click('.some-thing');
  await settled();
});
```
